reverting use babel's container APIs
These APIs cause babel to schedule further processing on our emitted code, which was my intent. But unfortunately, there are many other plugins that might decide to alter or remove these new nodes in between the time the work is scheduled and the time the work is done, and babel doesn't actually notice this. It will happily pass already-removed nodes to subsequent plugins, causing them to blow up.