Preserve order
Previously, we always inserted imports at the beginning of the file. This meant that the imports would end up in reverse order relative to the expressions in the file that caused those imports to be emitted.

This isn't strictly incorrect, but it's unintuitive and sometimes import order can matter, when modules have  ill-advised side-effects.