Reference-aware API
This is a breaking change to the API in order to guarantee that every identifier we create for you ends up with a valid reference in babel's binding system.

Previously, we created the bindings but it was up to callers to take the Identifiers we hand them and get them into the AST somewhere and then wire up the references. This is error prone.

Now, users must manipulate the AST within blocks of code wrapped by babel-import-util, ensuring that we can always install the correct references after your manipulations are done.

The primary methods are now:

 - replaceWith
 - insertAfter
 - insertBefore

These work just like the babel equivalents on NodePath, except they take a callback for building the replacement Node, and that callback has the ability to require imports.

There is also a lower-level `mutate` method that leaves the path mutations up to you, and requires you to return the modified path to babel-import-util.

Along with this change I refactored the tests to keep each test's example transform alongside the test itself. They should demonstrate many ways a plugin can use the library.